move logic to useSelectBuyAmount hook
